feat: add getSurveyFormStatus & SetCompletedForm to doppler legacy client
